+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Build 936 - 2026-01-18
+
+- automated build
+
 ## Build 935 - 2026-01-18
 
 - automated build